Updating dogbreeds.ipynb, the main report file
Uploaded an updated dogbreeds.ipynb file, wherein all import statements are moved to a cell at the start of the file. And the markdown write-up of the report is modified, made more concise, and adding an abstract. The write-up change is applied through the human-detection section to be continued with future changes.